Gather implementation for request parameters into W.T.C.Parameters.Internal
